Add remove method
for the distributed hashing use-case we have to occasionally remove nodes from the nodelist.
It obviously wouldn't affect prior object hashing, and a client would have to recalculate them.
But at least it is easier then nuking the hash, and recreating it from scratch.

The commit also adds a test, which finds a random key mapped to a node,
and then this node is removed, and then test validates that the key no
longer maps to the removed node.